fix: undo in slate
