+++ packages/core/src/components/Cell/Draggable/index.tsx
@@ -41,7 +41,8 @@ type Props = ComponetizedCell & {
   isInsertMode: boolean;
   isLayoutMode: boolean;
   node: { hover: string; inline: string };
-  children: React.ReactChildren;
+  // tslint:disable-next-line:no-any
+  children: any;
   className: string;
   name: string;
   dragType: string;
