+++ packages/core/src/reduxConnect.tsx
@@ -1,15 +1,18 @@
 import React from 'react';
-import { Provider, connect as reduxConnect } from 'react-redux';
+import { connect as reduxConnect, Provider } from 'react-redux';
 export const ReduxContext = React.createContext(null);
 
 export const ReduxProvider = ({ store, ...props }) => (
   <Provider store={store} context={ReduxContext} {...props} />
 );
 
-export const connect = (
-  mapStateToProps?: Function,
+// tslint:disable-next-line:no-any
+export const connect: any = (
+  // tslint:disable-next-line:no-any
+  mapStateToProps?: any,
   mapDispatchToProps?: Function | Object,
-  mergeProps?: Function,
+  // tslint:disable-next-line:no-any
+  mergeProps?: any,
   options: Object = {}
 ) =>
   reduxConnect(mapStateToProps, mapDispatchToProps, mergeProps, {
