+++ packages/plugins/content/slate/src/index.tsx
@@ -156,6 +156,7 @@ function plugin<TPlugins extends SlatePluginCollection = DefaultPlugins>(
     createInitialState: createInitialState,
     createInitialSlateState: createInitialState,
     htmlToSlate: htmlToSlate,
+    serialize: ({ slate }) => ({ slate }),
     unserialize: ({ slate, importFromHtml, ...rest }) => {
       if (importFromHtml) {
         return htmlToSlate(importFromHtml);
