+++ packages/plugins/content/slate/src/types/state.ts
@@ -1,5 +1,6 @@
-import { Node } from 'slate';
+import { Node, Range } from 'slate';
 
 export type SlateState = {
   slate: Node[];
+  selection?: Range;
 };
