+++ packages/ui/src/index.tsx
@@ -1,8 +1,8 @@
 // something is wrong with lerna, typescript and this import: import { lazyLoad } from '@react-page/core';
 import loadable from '@loadable/component';
+import { BottomToolbarProps } from './BottomToolbar/types';
 import { colorToString } from './ColorPicker/colorToString';
 import darkTheme from './ThemeProvider/DarkTheme';
-
 const Trash = loadable(() => import('./Trash/index'));
 const Toolbar = loadable(() => import('./Toolbar/index'));
 const DisplayModeToggle = loadable(() => import('./DisplayModeToggle/index'));
@@ -16,6 +16,7 @@ const Provider = loadable(() => import('./Provider'));
 
 export default EditorUI;
 export {
+  BottomToolbarProps,
   EditorUI,
   Provider,
   Trash,
